+++ app/db/models/__init__.py
@@ -1,14 +1,16 @@
 import os
 
 from app.db import get_db
+from app.db.models.categories import Category
 from app.db.models.dues_payments import DuesPayment
-from app.db.models.items import Item, Category
+from app.db.models.expense_accounts import ExpenseAccount
+from app.db.models.items import Item
 from app.db.models.member_donations import MemberDonation
 from app.db.models.member_due_payment import MemberDuesPayment
 from app.db.models.member_items import MemberItems
 from app.db.models.members import Member, MemberHistory
 from app.db.models.seller_items import SellerItems
-from app.db.models.sellers import Seller, ExpenseAccount
+from app.db.models.sellers import Seller
 
 
 def clear_db() -> None:
